requirements.txt
openai>=1.3.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
click>=8.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
tavily-python>=0.7.0

# Optional: LLMLingua for advanced compression (LLM-decided)
llmlingua>=0.2.0
transformers>=4.26.0
torch>=2.0.0
accelerate>=0.20.0
